packages/tenzo-agent/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
requests==2.31.0
psycopg2-binary==2.9.9
slowapi==0.1.9
PyJWT==2.8.0
bcrypt==4.1.3
web3==6.15.1
